ft_printf

+++ ft_write_long_int.c
@@ -6,7 +6,7 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/19 18:30:24 by cfatrane          #+#    #+#             */
-/*   Updated: 2016/12/27 18:36:23 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/06 19:49:53 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	ft_write_long_int(t_env *arg, va_list ap)
 {
 	int			len;
-	long int	nbr;
+	int	nbr;
 
 	len = 0;
 	if (arg->conv == 'D')
